refactoring to reuse query context cache implementation in ODBC
