+++ cpp/lib/QueryContextCache.cpp
@@ -7,70 +7,8 @@
 * Created on August 31, 2023, 4:05 PM
 */
 
-#include "QueryContextCache.hpp"
+#include "snowflake/QueryContextCache.hpp"
 #include "../logger/SFLogger.hpp"
-
-// wrapper functions for C
-extern "C" {
-
-  void qcc_initialize(SF_CONNECT * conn)
-  {
-    if (!conn || conn->qcc_disable || conn->qcc)
-    {
-      return;
-    }
-
-    conn->qcc = (void *) new Snowflake::Client::QueryContextCache(conn->qcc_capacity);
-  }
-
-  void qcc_set_capacity(SF_CONNECT * conn, uint64 capacity)
-  {
-    if (!conn || conn->qcc_disable)
-    {
-      return;
-    }
-    if (!conn->qcc)
-    {
-      qcc_initialize(conn);
-    }
-    static_cast<Snowflake::Client::QueryContextCache *>(conn->qcc)->setCapacity(capacity);
-  }
-
-  cJSON* qcc_serialize(SF_CONNECT * conn)
-  {
-    if (!conn || conn->qcc_disable || !conn->qcc)
-    {
-      return NULL;
-    }
-
-    return static_cast<Snowflake::Client::QueryContextCache *>(conn->qcc)->serializeQueryContext();
-  }
-
-  void qcc_deserialize(SF_CONNECT * conn, cJSON* query_context)
-  {
-    if (!conn || conn->qcc_disable)
-    {
-      return;
-    }
-    if (!conn->qcc)
-    {
-      qcc_initialize(conn);
-    }
-
-    static_cast<Snowflake::Client::QueryContextCache *>(conn->qcc)->deserializeQueryContext(query_context);
-  }
-
-  void qcc_terminate(SF_CONNECT * conn)
-  {
-    if (!conn || !conn->qcc)
-    {
-      return;
-    }
-
-    delete static_cast<Snowflake::Client::QueryContextCache *>(conn->qcc);
-  }
-
-} // extern "C"
 
 namespace Snowflake
 {
@@ -107,65 +45,19 @@ void QueryContextCache::setCapacity(size_t capacity)
   CXX_LOG_TRACE("QueryContextCache::setCapacity: set capacity from %d to %d", m_capacity, capacity);
 }
 
-void QueryContextCache::deserializeQueryContext(cJSON * data)
+void QueryContextCache::UpdateQueryContextCache(const std::vector<QueryContextElement>& entries)
 {
   std::lock_guard<std::recursive_mutex> guard(m_mutex);
   // Log existing cache entries
   logCacheEntries();
 
-  if ((!data) || (data->type != cJSON_Object))
+  for (auto entry = entries.begin(); entry != entries.end(); entry++)
   {
-    // Clear the cache
-    clearCache();
-    return;
+    merge(entry->id, entry->readTimestamp, entry->priority, entry->context);
   }
-
-  // Deserialize the entries. The first entry with priority is the main entry.
-  // Save all entries into one list to simplify the logic. An example JSON is:
-  // {
-  //   "entries": [
-  //    {
-  //     "id": 0,
-  //     "read_timestamp": 123456789,
-  //     "priority": 0,
-  //     "context": "base64 encoded context"
-  //    },
-  //     {
-  //       "id": 1,
-  //       "read_timestamp": 123456789,
-  //       "priority": 1,
-  //       "context": "base64 encoded context"
-  //     },
-  //     {
-  //       "id": 2,
-  //       "read_timestamp": 123456789,
-  //       "priority": 2,
-  //       "context": "base64 encoded context"
-  //     }
-  //   ]
-  cJSON * entries = snowflake_cJSON_GetObjectItem(data, QCC_ENTRIES_KEY);
-  if ((entries) && (entries->type == cJSON_Array))
-  {
-    int arraySize = snowflake_cJSON_GetArraySize(entries);
-    for (int i = 0; i < arraySize; i++)
-    {
-      QueryContextElement entry;
-      if (deserializeQueryContextElement(snowflake_cJSON_GetArrayItem(entries, i), entry))
-      {
-        merge(entry.id, entry.readTimestamp, entry.priority, entry.context);
-      }
-      else
-      {
-        CXX_LOG_TRACE("QueryContextCache::deserializeQueryContext: "
-                      "meets mismatch field type. Clear the QueryContextCache.");
-        clearCache();
-        return;
-      }
-    }
-    // after merging all entries, sync the internal priority map to priority map. Because of priority swicth from GS side,
-    // there could be priority key conflict if we directly operating on the priorityMap during a round of merge.
-    syncPriorityMap();
-  }
+  // after merging all entries, sync the internal priority map to priority map. Because of priority swicth from GS side,
+  // there could be priority key conflict if we directly operating on the priorityMap during a round of merge.
+  syncPriorityMap();
 
   // After merging all entries, truncate to capacity
   checkCacheCapacity();
@@ -174,78 +66,21 @@ void QueryContextCache::deserializeQueryContext(cJSON * data)
   logCacheEntries();
 }
 
-void QueryContextCache::deserializeQueryContextReq(cJSON * data)
+size_t QueryContextCache::GetQueryContextEntries(std::vector<QueryContextElement>& entries)
 {
   std::lock_guard<std::recursive_mutex> guard(m_mutex);
 
   // Log existing cache entries
   logCacheEntries();
 
-  if ((!data) || (data->type != cJSON_Object))
-  {
-    // Clear the cache
-    clearCache();
-    return;
-  }
+  entries.clear();
 
-  cJSON * entries = snowflake_cJSON_GetObjectItem(data, QCC_ENTRIES_KEY);
-  int arraySize = snowflake_cJSON_GetArraySize(entries);
-  if ((entries) && (entries->type == cJSON_Array))
-  {
-    for (int i = 0; i < arraySize; i++)
-    {
-      QueryContextElement entry;
-      if (deserializeQueryContextElementReq(snowflake_cJSON_GetArrayItem(entries, i), entry))
-      {
-        merge(entry.id, entry.readTimestamp, entry.priority, entry.context);
-      }
-      else
-      {
-        CXX_LOG_TRACE("QueryContextCache::deserializeQueryContextReq: "
-          "meets mismatch field type. Clear the QueryContextCache.");
-        clearCache();
-        return;
-      }
-    }
-    // after merging all entries, sync the internal priority map to priority map. Because of priority swicth from GS side,
-    // there could be priority key conflict if we directly operating on the priorityMap during a round of merge.
-    syncPriorityMap();
-  }
-
-  // After merging all entries, truncate to capacity
-  checkCacheCapacity();
-
-  // Log existing cache entries
-  logCacheEntries();
-}
-
-cJSON *  QueryContextCache::serializeQueryContext()
-{
-  std::lock_guard<std::recursive_mutex> guard(m_mutex);
-
-  // Log existing cache entries
-  logCacheEntries();
-
-  cJSON * queryContext = snowflake_cJSON_CreateObject();
-  cJSON * entries = snowflake_cJSON_CreateArray();
   for (auto itr = m_cacheSet.begin(); itr != m_cacheSet.end(); itr++)
   {
-    cJSON * entry = snowflake_cJSON_CreateObject();
-    snowflake_cJSON_AddUint64ToObject(entry, QCC_ID_KEY, itr->id);
-    snowflake_cJSON_AddUint64ToObject(entry, QCC_PRIORITY_KEY, itr->priority);
-    snowflake_cJSON_AddUint64ToObject(entry, QCC_TIMESTAMP_KEY, itr->readTimestamp);
-    cJSON * context = snowflake_cJSON_CreateObject();
-    if (!itr->context.empty())
-    {
-      snowflake_cJSON_AddStringToObject(context, QCC_CONTEXT_VALUE_KEY, itr->context.c_str());
-    }
-    snowflake_cJSON_AddItemToObject(entry, QCC_CONTEXT_KEY, context);
-    snowflake_cJSON_AddItemToArray(entries, entry);
+    entries.emplace_back(*itr);
   }
 
-  snowflake_cJSON_AddItemToObject(queryContext, QCC_ENTRIES_KEY, entries);
-
-  return queryContext;
+  return entries.size();
 }
 
 void QueryContextCache::merge(uint64 id, uint64 readTimestamp,
@@ -350,124 +185,6 @@ size_t QueryContextCache::getElements(std::vector<uint64>& ids,
 }
 
 // Private =====================================================================
-bool QueryContextCache::deserializeQueryContextElement(cJSON * entryNode,
-                                                       QueryContextElement & contextElement)
-{
-  cJSON * id = snowflake_cJSON_GetObjectItem(entryNode, QCC_ID_KEY);
-  if (id && (id->type == cJSON_Number))
-  {
-    contextElement.id = snowflake_cJSON_GetUint64Value(id);
-  }
-  else
-  {
-    CXX_LOG_WARN("QueryContextCache::deserializeQueryContextElement: `id` field is not integer type");
-    return false;
-  }
-
-  cJSON * timestamp = snowflake_cJSON_GetObjectItem(entryNode, QCC_TIMESTAMP_KEY);
-  if (timestamp && (timestamp->type == cJSON_Number))
-  {
-    contextElement.readTimestamp = snowflake_cJSON_GetUint64Value(timestamp);
-  }
-  else
-  {
-    CXX_LOG_WARN("QueryContextCache::deserializeQueryContextElement: `timestamp` field is not integer type");
-    return false;
-  }
-
-  cJSON * priority = snowflake_cJSON_GetObjectItem(entryNode, QCC_PRIORITY_KEY);
-  if (priority && (priority->type == cJSON_Number))
-  {
-    contextElement.priority = snowflake_cJSON_GetUint64Value(priority);
-  }
-  else
-  {
-    CXX_LOG_WARN("QueryContextCache::deserializeQueryContextElement: `priority` field is not integer type");
-    return false;
-  }
-
-  cJSON * context = snowflake_cJSON_GetObjectItem(entryNode, QCC_CONTEXT_KEY);
-  if (!context || context->type == cJSON_NULL)
-  {
-    return true;
-  }
-  if (context->type == cJSON_String)
-  {
-    contextElement.context = snowflake_cJSON_GetStringValue(context);
-  }
-  else
-  {
-    CXX_LOG_WARN("QueryContextCache::deserializeQueryContextElement: `context` field is not string type");
-    return false;
-  }
-
-  return true;
-}
-
-bool QueryContextCache::deserializeQueryContextElementReq(cJSON * entryNode,
-                                                          QueryContextElement & contextElement)
-{
-  cJSON * id = snowflake_cJSON_GetObjectItem(entryNode, QCC_ID_KEY);
-  if (id && (id->type == cJSON_Number))
-  {
-    contextElement.id = snowflake_cJSON_GetUint64Value(id);
-  }
-  else
-  {
-    CXX_LOG_WARN("QueryContextCache::deserializeQueryContextElement: `id` field is not integer type");
-    return false;
-  }
-
-  cJSON * timestamp = snowflake_cJSON_GetObjectItem(entryNode, QCC_TIMESTAMP_KEY);
-  if (timestamp && (timestamp->type == cJSON_Number))
-  {
-    contextElement.readTimestamp = snowflake_cJSON_GetUint64Value(timestamp);
-  }
-  else
-  {
-    CXX_LOG_WARN("QueryContextCache::deserializeQueryContextElement: `timestamp` field is not integer type");
-    return false;
-  }
-
-  cJSON * priority = snowflake_cJSON_GetObjectItem(entryNode, QCC_PRIORITY_KEY);
-  if (priority && (priority->type == cJSON_Number))
-  {
-    contextElement.priority = snowflake_cJSON_GetUint64Value(priority);
-  }
-  else
-  {
-    CXX_LOG_WARN("QueryContextCache::deserializeQueryContextElement: `priority` field is not integer type");
-    return false;
-  }
-
-  cJSON * context = snowflake_cJSON_GetObjectItem(entryNode, QCC_CONTEXT_KEY);
-  if (!context || context->type == cJSON_NULL)
-  {
-    return true;
-  }
-  if (context->type != cJSON_Object)
-  {
-    return false;
-  }
-
-  cJSON * contextValue = snowflake_cJSON_GetObjectItem(context, QCC_CONTEXT_VALUE_KEY);
-  if (!contextValue || contextValue->type == cJSON_NULL)
-  {
-    return true;
-  }
-
-  if (contextValue->type == cJSON_String)
-  {
-    contextElement.context = snowflake_cJSON_GetStringValue(contextValue);
-  }
-  else
-  {
-    return false;
-  }
-
-  return true;
-}
-
 void QueryContextCache::addQCE(const QueryContextElement& qce)
 {
   m_idMap[qce.id] = qce;
